Remove schemabuilder interfaces
