+++ src/main/java/com/hivemq/adapter/sdk/api/schema/PropertySchemaBuilder.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023-present HiveMQ GmbH
+ * Copyright 2019-present HiveMQ GmbH
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -25,135 +25,195 @@ import org.jetbrains.annotations.NotNull;
  *
  * @param <P> the parent builder type of the containing {@link ObjectSchemaBuilder}.
  */
-public interface PropertySchemaBuilder<P> {
+public final class PropertySchemaBuilder<P> extends AbstractSchemaBuilder<PropertySchemaBuilder<P>> {
+
+    private final ObjectSchemaBuilder<P> parentObject;
+    private final String key;
+
+    PropertySchemaBuilder(final ObjectSchemaBuilder<P> parentObject, final String key) {
+        super("property '" + key + "'");
+        this.parentObject = parentObject;
+        this.key = key;
+    }
 
     // ── Property-specific ────────────────────────────────────────────────────
 
     /**
      * Mark this property as required. Equivalent to {@code required(true)}.
      */
-    @NotNull PropertySchemaBuilder<P> required();
+    public @NotNull PropertySchemaBuilder<P> required() {
+        return required(true);
+    }
 
     /**
      * Set whether this property must be present in the object.
      */
-    @NotNull PropertySchemaBuilder<P> required(boolean required);
+    public @NotNull PropertySchemaBuilder<P> required(final boolean required) {
+        if (required) {
+            parentObject.required.add(key);
+        } else {
+            parentObject.required.remove(key);
+        }
+        return self();
+    }
 
     // ── Structure-defining calls ─────────────────────────────────────────────
 
     /**
      * No type restriction; any value is valid for this property.
      */
-    @NotNull PropertySchemaBuilder<P> any();
+    public @NotNull PropertySchemaBuilder<P> any() {
+        return doAny();
+    }
 
     /**
      * A single primitive type for this property.
      */
-    @NotNull PropertySchemaBuilder<P> scalar(@NotNull ScalarType type);
+    public @NotNull PropertySchemaBuilder<P> scalar(final @NotNull ScalarType type) {
+        return doScalar(type);
+    }
 
     /**
-     * Begin a nested object as this property's value impl.
+     * Begin a nested object as this property's value schema.
      */
-    @NotNull ObjectSchemaBuilder<PropertySchemaBuilder<P>> startObject();
+    public @NotNull ObjectSchemaBuilder<PropertySchemaBuilder<P>> startObject() {
+        return doStartObject(this);
+    }
 
     /**
-     * Begin a nested array as this property's value impl.
+     * Begin a nested array as this property's value schema.
      */
-    @NotNull ItemSchemaBuilder<PropertySchemaBuilder<P>> startArray();
+    public @NotNull ItemSchemaBuilder<PropertySchemaBuilder<P>> startArray() {
+        return doStartArray(this);
+    }
 
     /**
-     * Use an already-complete {@link Schema} as this property's value impl.
+     * Use an already-complete {@link Schema} as this property's value schema.
      * Annotation methods must not be called after this.
      */
-    @NotNull PropertySchemaBuilder<P> schema(@NotNull Schema schema);
+    public @NotNull PropertySchemaBuilder<P> schema(final @NotNull Schema schema) {
+        return doSchema(schema);
+    }
 
-    // ── Structure-defining flag ──────────────────────────────────────────────
+    // ── Nullable flag ────────────────────────────────────────────────────────
 
     /**
      * Mark this property's value as nullable. Equivalent to {@code nullable(true)}.
      */
-    @NotNull PropertySchemaBuilder<P> nullable();
+    public @NotNull PropertySchemaBuilder<P> nullable() {
+        return doNullable(true);
+    }
 
     /**
      * Set whether {@code null} is a valid value for this property. Default is {@code false}.
      */
-    @NotNull PropertySchemaBuilder<P> nullable(boolean nullable);
+    public @NotNull PropertySchemaBuilder<P> nullable(final boolean nullable) {
+        return doNullable(nullable);
+    }
 
     // ── Range constraints ────────────────────────────────────────────────────
 
     /**
      * Set the inclusive lower bound for numeric scalar types.
      */
-    @NotNull PropertySchemaBuilder<P> minimum(long minimum);
+    public @NotNull PropertySchemaBuilder<P> minimum(final long minimum) {
+        return doMinimum(minimum);
+    }
 
     /**
      * Set the inclusive upper bound for numeric scalar types.
      */
-    @NotNull PropertySchemaBuilder<P> maximum(long maximum);
+    public @NotNull PropertySchemaBuilder<P> maximum(final long maximum) {
+        return doMaximum(maximum);
+    }
 
     /**
      * Set the inclusive lower bound for floating-point scalar types.
      */
-    @NotNull PropertySchemaBuilder<P> minimum(double minimum);
+    public @NotNull PropertySchemaBuilder<P> minimum(final double minimum) {
+        return doMinimum(minimum);
+    }
 
     /**
      * Set the inclusive upper bound for floating-point scalar types.
      */
-    @NotNull PropertySchemaBuilder<P> maximum(double maximum);
+    public @NotNull PropertySchemaBuilder<P> maximum(final double maximum) {
+        return doMaximum(maximum);
+    }
 
     // ── Annotations ──────────────────────────────────────────────────────────
 
     /**
      * Set a short human-readable label for this property.
      */
-    @NotNull PropertySchemaBuilder<P> title(@NotNull String title);
+    public @NotNull PropertySchemaBuilder<P> title(final @NotNull String title) {
+        return doTitle(title);
+    }
 
     /**
      * Set a longer human-readable explanation for this property.
      */
-    @NotNull PropertySchemaBuilder<P> description(@NotNull String description);
+    public @NotNull PropertySchemaBuilder<P> description(final @NotNull String description) {
+        return doDescription(description);
+    }
 
     /**
      * Mark this property as readable. Equivalent to {@code readable(true)}.
      */
-    @NotNull PropertySchemaBuilder<P> readable();
+    public @NotNull PropertySchemaBuilder<P> readable() {
+        return doReadable(true);
+    }
 
     /**
      * Set whether clients may read this property's value. Default is {@code true}.
      */
-    @NotNull PropertySchemaBuilder<P> readable(boolean readable);
+    public @NotNull PropertySchemaBuilder<P> readable(final boolean readable) {
+        return doReadable(readable);
+    }
 
     /**
      * Mark this property as writable. Equivalent to {@code writable(true)}.
      */
-    @NotNull PropertySchemaBuilder<P> writable();
+    public @NotNull PropertySchemaBuilder<P> writable() {
+        return doWritable(true);
+    }
 
     /**
      * Set whether clients may write this property's value. Default is {@code false}.
      */
-    @NotNull PropertySchemaBuilder<P> writable(boolean writable);
+    public @NotNull PropertySchemaBuilder<P> writable(final boolean writable) {
+        return doWritable(writable);
+    }
 
     // ── Navigation ───────────────────────────────────────────────────────────
 
     /**
      * Return to the containing {@link ObjectSchemaBuilder}.
      */
-    @NotNull ObjectSchemaBuilder<P> endProperty();
+    public @NotNull ObjectSchemaBuilder<P> endProperty() {
+        return parentObject;
+    }
 
     // ── Shorthands (delegate through endProperty()) ──────────────────────────
 
     /**
      * Shorthand for {@code endProperty().property(key)}.
      */
-    @NotNull PropertySchemaBuilder<P> property(@NotNull String key);
+    public @NotNull PropertySchemaBuilder<P> property(final @NotNull String key) {
+        return endProperty().property(key);
+    }
 
     /**
      * Shorthand for {@code endProperty().additionalProperties(additionalProperties)}.
      */
-    @NotNull ObjectSchemaBuilder<P> additionalProperties(boolean additionalProperties);
+    public @NotNull ObjectSchemaBuilder<P> additionalProperties(final boolean additionalProperties) {
+        return endProperty().additionalProperties(additionalProperties);
+    }
 
     /**
      * Shorthand for {@code endProperty().endObject()}.
      */
-    @NotNull P endObject();
+    public @NotNull P endObject() {
+        return endProperty().endObject();
+    }
 }
