+++ src/main/java/com/hivemq/adapter/sdk/api/schema/SchemaAnnotations.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hivemq.adapter.sdk.api.schema.impl;
+package com.hivemq.adapter.sdk.api.schema;
 
 import org.jetbrains.annotations.Nullable;
